requirements.txt
cairocffi==1.0.2
CairoSVG==2.4.0
cffi==1.12.3
cssselect2==0.2.1
defusedxml==0.6.0
Django==2.2.3
django-crispy-forms==1.7.2
django-mathfilters==0.4.0
django-mysql==3.0.0.post1
html5lib==1.0.1
mysqlclient==1.4.2.post1
Pillow==6.1.0
pycparser==2.19
PyMySQL==0.9.3
Pyphen==0.9.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
tinycss2==1.0.2
WeasyPrint==48
webencodings==0.5.1
xlwt==1.3.0
